refactor: change AvaliacaoEntity.avaliacaoFinal type from BigDecimal to Double

+++ src/main/java/cv/inps/rh/shared/infrastructure/persistence/entity/AvaliacaoEntity.java
@@ -52,7 +52,7 @@ public class AvaliacaoEntity extends AuditEntity {
   private String estado;                       // 'A' | 'P' | 'C'
 
   @Column(name = "AVALIACAO_FINAL")
-  private BigDecimal avaliacaoFinal;
+  private Double avaliacaoFinal;
 
   @Column(name = "PESO_COMPORTAMENTAIS")
   private BigDecimal pesoComportamentais;
